backend/requirements.txt
# FastAPI 및 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Socket.IO (Flask-SocketIO 호환)
python-socketio==5.12.0

# WebSocket
websockets==12.0

# HTTP 클라이언트
requests==2.31.0

# OpenAI (최신 안정 버전)
openai>=1.3.0,<2.0.0

# 환경 변수
python-dotenv==1.0.0

# 데이터 모델
pydantic==2.5.0

# 오디오 처리 (프론트엔드에서 처리하므로 선택사항)
# pyaudio==0.2.11
# numpy==1.24.3